Use custom `traffic-root` layer
With the update to Mapbox GL JS 3 and the introduction of the new Mapbox
Standard Style, the layers with `source-layer: 'road'` seem to have
disappeared. The workaround I've found is to add a custom layer with an
ID of `traffic-root` just before initializing the Mapbox Traffic Layer.
This change directly looks for that new layer.